+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project are documented in this file. This project
 adheres to [Semantic Versioning](https://semver.org/).
 
 ## 0.0.5 (XX-XX-2025)
+* Increased min python version to 3.10 (which is a requirement for some scipy functions)
+* Added a new capability to do the Apex coordinate transformation without ApexPy
+* Added unit tests for Apex_library
 * Corrected foEs interpolation method and updated test data and example figures.
 * Deprecated `output_quartiles` argument in `read_ccir_ursi_coeff()` and replaced it with `output_deciles` (functionality is unchanged).
 * Some typo fixes